removed typecasts and minor stuff.

+++ src/edu/hm/fuberg/se2/android/robotz/control/PillChecker.java
@@ -17,7 +17,7 @@ import edu.hm.fuberg.se2.android.robotz.data.Arena;
  * Class for defining various check methods concerning the ultimate pill for invincibility.
  * @author Stephanie Ehrenberg
  * @author Robert Fuess
- * @version 2013-05-27
+ * @version 2013-06-15
  */
 class PillChecker {
 
@@ -46,7 +46,7 @@ class PillChecker {
 		if (robotzData.getInvinciblePill() == null) {
 
 			final Random random = new Random();
-			final int probability = random.nextInt((int)configurator.getRandomPill());
+			final int probability = random.nextInt(configurator.getRandomPill());
 
 			if (probability == 0) {
 				createPossiblePill(random);
@@ -69,7 +69,7 @@ class PillChecker {
 
 			if (!invinciblePillOnItem(xCoord, yCoord)) {
 
-				robotzData.setInvinciblePill(xCoord, yCoord, (int)configurator.getDurationPill());
+				robotzData.setInvinciblePill(xCoord, yCoord, configurator.getDurationPill());
 				noFreeSlot = false;
 			}
 		}
